feat: combine cli query server into a signle file
