--- x/alliance/client/cli/query_alliances.go
@@ -1,74 +1,0 @@
-package cli
-
-import (
-	"strconv"
-
-	"alliance/x/alliance/types"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/spf13/cobra"
-)
-
-var _ = strconv.Itoa(0)
-
-func CmdQueryAlliances() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "alliances",
-		Short: "Query paginated alliances",
-		Args:  cobra.ExactArgs(0),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-
-			clientCtx, err := client.GetClientQueryContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			params := &types.QueryAlliancesRequest{}
-
-			res, err := queryClient.Alliances(cmd.Context(), params)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
-
-func CmdQueryAlliance() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "alliance denom",
-		Short: "Query a specific alliance by denom",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			denom := args[0]
-
-			ctx, err := client.GetClientQueryContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			query := types.NewQueryClient(ctx)
-
-			params := &types.QueryAllianceRequest{Denom: denom}
-
-			res, err := query.Alliance(cmd.Context(), params)
-			if err != nil {
-				return err
-			}
-
-			return ctx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
