--- x/alliance/client/cli/query_rewards.go
@@ -1,42 +1,0 @@
-package cli
-
-import (
-	"context"
-
-	"alliance/x/alliance/types"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/spf13/cobra"
-)
-
-func CmdQueryRewards() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "rewards delegator_addr validator_addr denom",
-		Short: "Query module parameters",
-		Args:  cobra.ExactArgs(3),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			delegatorAddr := args[0]
-			validatorAddr := args[1]
-			denom := args[2]
-			ctx := client.GetClientContextFromCmd(cmd)
-			query := types.NewQueryClient(ctx)
-			params := &types.QueryAllianceDelegationRewardsRequest{
-				DelegatorAddr: delegatorAddr,
-				ValidatorAddr: validatorAddr,
-				Denom:         denom,
-			}
-
-			res, err := query.AllianceDelegationRewards(context.Background(), params)
-			if err != nil {
-				return err
-			}
-
-			return ctx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
